UPDATE: merge html_template to send_mail
