+++ server/src/utils/send_mail.ts
@@ -1,7 +1,17 @@
 import mailer from "@/configs/mailer";
 import env from "@/configs/env";
-import { renderTemplate } from "./html_template";
+import * as handlebars from 'handlebars';
+import * as fs from 'fs';
+import * as path from 'path';
 import { IQueryableUser } from "@/types/auth";
+
+export const renderTemplate = (templatePath: string, replacements: object): string => {
+    const filePath = path.join(path.resolve(), templatePath);
+    const source = fs.readFileSync(filePath, 'utf-8').toString();
+    const template = handlebars.compile(source);
+
+    return template(replacements);
+}
 
 export async function sendMail(to: string, subject: string, body: any) {
     await mailer.sendMail({
